requirements.txt
requests
tweepy
nltk
textblob
python-dotenv
discord